+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Added:
 - Added default dungeon config option
 - Added boss nbt option
+- Added zombification tag
+- Added min group size option
 ### Fixed:
 - 
 ### Changed: